backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-socketio==5.11.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiosmtplib==3.0.1
python-multipart==0.0.9
supabase==2.9.1
pydantic-settings==2.5.2
python-dotenv==1.0.1
Pillow==10.4.0
PyPDF2==3.0.1
pydub==0.25.1
